Archiv változás

+++ WIWEdetector/Archive.vb
@@ -7,15 +7,24 @@ Module Archive
         Dim strTimeStamp As String
         Dim newFileName As String
         Dim path As String = Directory.GetCurrentDirectory()
+        Dim res As Integer
 
-        strTimeStamp = dt.ToString
-        strTimeStamp = strTimeStamp.Replace(".", "")
-        strTimeStamp = strTimeStamp.Replace(":", "")
-        strTimeStamp = strTimeStamp.Replace(" ", "_")
+        res = MessageBox.Show("A meglévő adatbázis archiválásra kerül. Üres adatbázist akarsz használni?", "FIGYELEM!", MessageBoxButtons.YesNoCancel)
 
-        newFileName = path & "\Arch\WIWEdevices_" & strTimeStamp & ".s3db"
+        If res = DialogResult.Yes Or res = DialogResult.No Then
 
-        Console.WriteLine(newFileName)
-        File.Copy("WIWEdevices.s3db", newFileName, True)
+            strTimeStamp = dt.ToString
+            strTimeStamp = strTimeStamp.Replace(".", "")
+            strTimeStamp = strTimeStamp.Replace(":", "")
+            strTimeStamp = strTimeStamp.Replace(" ", "_")
+
+            newFileName = path & "\Arch\WIWEdevices_" & strTimeStamp & ".s3db"
+
+            Console.WriteLine(newFileName)
+            File.Copy("WIWEdevices.s3db", newFileName, True)
+            If res = DialogResult.Yes Then
+                File.Copy(path & "\Lib\WIWEdevices.s3db", path & "\WIWEdevices.s3db", True)
+            End If
+        End If
     End Sub
 End Module
